Added stellar pop obs and changed plot names
To avoid conflicting names when we compare among simulated catalogues and with observations, I changed
the ending of the names of the plots in which we compare to observations to explicitely reflect that.

I also added the observations of van de Sande et al. (2019) to compare the predicted stellar velocity
dispersion and stella age. The caveat is that those obs are r-band weighted rather than stellar mass
weighted.